refactor(utxo-core): split Output and descriptor Output
Issue: BTC-1826
