+++ modules/utxo-core/src/descriptor/index.ts
@@ -3,3 +3,6 @@ export * from './address';
 export * from './DescriptorMap';
 export * from './Output';
 export * from './VirtualSize';
+
+/** @deprecated - import from @bitgo/utxo-core directly instead */
+export * from '../Output';
