+++ modules/utxo-core/src/index.ts
@@ -2,3 +2,4 @@ export * as bip65 from './bip65';
 export * as descriptor from './descriptor';
 export * as testutil from './testutil';
 export * from './dustThreshold';
+export * from './Output';
